+++ Model/Service/PayerAuth/Request/AuthenticationRequest.php
@@ -22,6 +22,7 @@
 namespace ParadoxLabs\CyberSource\Model\Service\PayerAuth\Request;
 
 use Magento\Framework\Exception\InputException;
+use ParadoxLabs\CyberSource\Model\Service\UnifiedCheckout\LegacyTokenTrait;
 use ParadoxLabs\CyberSource\Model\Service\UnifiedCheckout\Request\FilterEmptyTrait;
 
 /**
@@ -41,6 +42,7 @@ use ParadoxLabs\CyberSource\Model\Service\UnifiedCheckout\Request\FilterEmptyTra
 class AuthenticationRequest
 {
     use FilterEmptyTrait;
+    use LegacyTokenTrait;
 
     /**
      * Browser fields the request must always carry; thin data is a silent 3DS bypass.
@@ -407,10 +409,12 @@ class AuthenticationRequest
 
         if ($this->hasTransientToken()) {
             $request['tokenInformation'] = ['transientTokenJwt' => $this->transientToken];
+        } elseif (!empty($this->card)) {
+            $request['paymentInformation'] = ['card' => $this->card];
+        } elseif ($this->isLegacyToken($this->paymentInstrumentId)) {
+            $request['paymentInformation'] = ['customer' => ['customerId' => $this->paymentInstrumentId]];
         } else {
-            $request['paymentInformation'] = !empty($this->card)
-                ? ['card' => $this->card]
-                : ['paymentInstrument' => ['id' => $this->paymentInstrumentId]];
+            $request['paymentInformation'] = ['paymentInstrument' => ['id' => $this->paymentInstrumentId]];
         }
 
         return $request;